gluco twin/requirements.txt
numpy>=1.21
flask>=2.0
pyserial>=3.5
vosk>=0.3.45
pyttsx3>=2.90
sounddevice>=0.4.5
